[WIP] Don't merge. It doesn't work.
This is a work in progress. It's intention was to move to the
use of components so that measuring.device was implemented using
them. The previous commit to this one is the one that works,
but doesn't use components (but simple 2nd-type of inheritance).
