+++ stock_measuring_device/models/__init__.py
@@ -2,3 +2,4 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
 from . import measuring_device
 from . import product_packaging
+from . import stock_warehouse
