+++ stock_measuring_device/models/product_packaging.py
@@ -1,34 +1,32 @@
 # Copyright 2021 Camptocamp SA
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
-from odoo import api, models
+from odoo import api, fields, models
 
 
 class ProductPackaging(models.Model):
     _inherit = "product.packaging"
 
-    @api.model
-    def _acquire_measuring_device(self, device):
-        """Lock the measuring device for other product packaging
+    measuring_device_id = fields.Many2one(
+        "measuring.device",
+        copy=False,
+        string="Measuring device which will scan the package",
+        help="Technical field set when an operator uses the device "
+        "to scan this package",
+    )
 
-        Intended to be extended. Each package defines a set of specific
-        measuring devices and depending on device._name it uses one
-        or other implementation by calling super()"""
-        pass
+    @api.model
+    def _acquire_measuring_device(self):
+        """Lock the measuring device for other product packaging"""
+        self.env.cr.execute(
+            "SELECT measuring_device_id FROM product_packaging FOR UPDATE"
+        )
 
     def _assign_measuring_device(self, device):
-        """Assign the measuring device to the current product packaging
-
-        Intended to be extended. Each package defines a set of specific
-        measuring devices and depending on device._name it uses one
-        or other implementation by calling super()"""
+        """Assign the measuring device to the current product packaging"""
         self.ensure_one()
-        pass
+        self.measuring_device_id = device
 
-    def _release_measuring_device(self, device):
-        """Free the measuring device from the current product packaging
-
-        Intended to be extended. Each package defines a set of specific
-        measuring devices and depending on device._name it uses one
-        or other implementation by calling super()"""
+    def _release_measuring_device(self):
+        """Free the measuring device from the current product packaging"""
         self.ensure_one()
-        pass
+        self.measuring_device_id = False
