+++ stock_measuring_device/models/stock_warehouse.py
@@ -7,6 +7,6 @@ from odoo import fields, models
 class StockWarehouse(models.Model):
     _inherit = "stock.warehouse"
 
-    zippcube_device_ids = fields.One2many(
-        "zippcube.device", "warehouse_id", string="Zippcube Devices"
+    measuring_device_ids = fields.One2many(
+        "measuring.device", "warehouse_id", string="Measuring Devices"
     )
