+++ stock_measuring_device_zippcube/components/__init__.py
@@ -1,5 +1,3 @@
 # Copyright 2021 Camptocamp SA
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
-from . import components
-from . import models
-from . import wizard
+from . import zippcube_device_component
